Simplified the noise terms
Now just add a complex number with gaussian amplitude and random phase
to plane wave. (Does it matter that it is just a scalar?

Also, started using amp = 10.0 and darkcts = 0 (to ignore them). Seems
to work correctly, as shown in simpler_noise.png. The averages for the
real and imaginary part make sense for the given signals.